challenge-4/backend/chatbot/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.17.2
pandas==2.2.2
google-genai==1.20.0
Flask==2.3.3
db-dtypes==1.1.1
